Add a higher-order component for prop conversion
Because props are immutable inside a component, there were a couple
things I wanted to do, but couldn't:

* Convert strings to integers for props like `duration`, because it
is easy to imagine people passing `duration="500"` instead of
`duration={500}`.

* Deal with the React.Children.toArray ugliness in one place, instead
of having to convert it in 3 different places.

By moving this (as well as propType validation and defaultProps) to
a higher-order component, I abstract away this bit of crucial but
uninteresting code, so that FlipMove.js concerns itself with its
actual job.

Still to come: Adding further propType validation (making sure that
`duration` is a value between 1 and 100,000, for example)